backend/pyproject.toml
[tool.poetry]
name = "."
version = "0.1.0"
description = ""
authors = ["zeph <zephyrr2722@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^3.2"
psycopg2-binary = "^2.9.3"
django-environ = "^0.8.1"
djangorestframework = "^3.13.1"
djangorestframework-simplejwt = "^5.0.0"
gunicorn = "^20.1.0"
firebase-admin = "^5.2.0"
fcm-django = "^1.0.7"
django-cors-headers = "^3.10.0"
drf-yasg = "^1.20.0"
django-hosts = "^5.0"
Pillow = "^9.0.0"
boto3 = "^1.20.26"
django-filter = "^21.1"
django-datatables-view = "^1.19.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
